backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
gunicorn==21.2.0
